requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
boto3==1.14.2
botocore==1.17.2
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-fontawesome==1.0
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-js-asset==1.2.2
-e git+https://github.com/gbozee/django-paystack.git@cfdf950e9f7fdc10fe6d4118b3a3327a68db68c9#egg=django_paystack
django-phone-field==1.8.0
django-sendgrid-v5==0.8.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.9
jmespath==0.10.0
Paperboy==1.0.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-async==0.6.2
rfc3986==1.4.0
s3transfer==0.3.3
sendgrid==6.3.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
